+++ pyproject.toml
@@ -19,15 +19,12 @@ license = { file = "LICENSE" }
 dependencies = [
   "aiocache>=0.12.2",
   "aio-eapi>=0.6.3",
-  "click>=8.1.6",
-  "click-help-colors>=0.9",
   "cvprac>=1.3.1",
   "pydantic>=2.6.1",
   "pydantic-extra-types>=2.3.0",
   "eval-type-backport>=0.1.3",  # Support newer typing features in older Python versions (required until Python 3.9 support is removed)
   "PyYAML>=6.0",
   "requests>=2.31.0",
-  "rich>=13.5.2",
   "asyncssh>=2.13.2",
   "Jinja2>=3.1.2",
 ]
@@ -53,6 +50,11 @@ classifiers = [
 requires-python = ">=3.9"
 
 [project.optional-dependencies]
+cli = [
+  "click~=8.1.6",
+  "click-help-colors>=0.9",
+  "rich>=13.5.2,<14",
+]
 dev = [
   "bumpver>=2023.1129",
   "codespell~=2.2.6",
@@ -382,6 +384,9 @@ runtime-evaluated-base-classes = ["pydantic.BaseModel", "anta.models.AntaTest.In
 ]
 "anta/cli/exec/utils.py" = [
   "SLF001",  # TODO: some private members, lets try to fix
+]
+"anta/cli/__init__.py" = [
+  "T201",  # Allow print statements
 ]
 "anta/cli/*" = [
   "PLR0913",  # Allow more than 5 input arguments in CLI functions
