+++ tests/units/cli/get/test__init__.py
@@ -7,7 +7,7 @@ from __future__ import annotations
 
 from typing import TYPE_CHECKING
 
-from anta.cli import anta
+from anta.cli._main import anta
 from anta.cli.utils import ExitCode
 
 if TYPE_CHECKING:
